% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverallFunction.R
\name{HeterogeneousTEpanel}
\alias{HeterogeneousTEpanel}
\title{Estimate causal effects of a policy with staggered adoption of treatment}
\usage{
HeterogeneousTEpanel(
  outcomes,
  treatments,
  covariates = NULL,
  nTimesOut = 10,
  nScans = 5000,
  nBurn = 2000,
  thin = 6,
  zeroMat = NULL,
  smoothEffects = TRUE,
  newX = NULL,
  priorA = 0.33,
  priorB = 3.3e-06
)
}
\arguments{
\item{outcomes}{An N by T matrix of outcomes for each unit over T time periods}

\item{treatments}{An N by T matrix of binary indicators indicating treatment
status for each unit at each point in time}

\item{covariates}{An N by Q data frame of covariate values for each unit}

\item{nScans}{The number of MCMC scans to run}

\item{nBurn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan
that is kept}

\item{zeroMat}{An optional matrix of indices of entries of the inverse
covariance matrix of the time series to be set to zero. 
If NULL, then no elements will be forced to zero. You do
not need to specify both the (i,j) and (j,i) elements
as the solution is necessarily symmetric. For more details
see the helpfile for glasso().}

\item{smoothEffects}{Whether or not to smooth treatment effects over time}

\item{newX}{An M by Q matrix of covariate values to estimate the 
treatment effect at. The default is NULL.}

\item{priorA}{First hyperprior for inverse gamma prior distribution
on variance components for the time series state parameters.
We don't recommend changing this parameter, but generally
larger values lead to smoother time series forecasts.}

\item{priorB}{Second hyperprior for inverse gamma prior distribution
on variance components for the time series state parameters.
We don't recommend changing this parameter, but generally
smaller values lead to smoother time series forecasts.}
}
\value{
Description
}
\description{
Estimate causal effects of a policy with staggered adoption of treatment
}
