% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotHeterogeneousEffects}
\alias{PlotHeterogeneousEffects}
\title{Plot heterogeneous effects of a policy with staggered adoption of treatment}
\usage{
PlotHeterogeneousEffects(modelFit)
}
\arguments{
\item{modelFit}{The result of a call to HeterogeneousTEpanel()}
}
\value{
Estimates and 95% credible intervals for the effect of each
        covariate considered.
}
\description{
Plot heterogeneous effects of a policy with staggered adoption of treatment
}
