% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{PlotMarginalEffects}
\alias{PlotMarginalEffects}
\title{Plot marginal effects of a policy with staggered adoption of treatment}
\usage{
PlotMarginalEffects(modelFit)
}
\arguments{
\item{modelFit}{The result of a call to HeterogeneousTEpanel()}
}
\value{
Estimates and 95% credible intervals for marginal estimands 
        at each time point considered.
}
\description{
Plot marginal effects of a policy with staggered adoption of treatment
}
